banao_ai/task2/requirements.txt
transformers 
torch 
peft 
accelerate 
langgraph 
huggingface_hub
python-dotenv